+++ CHANGES.md
@@ -4,6 +4,7 @@ CHANGES
 ## 2.4.2-RC
 
 * Improvement: Ensure to set absolute url from root when attaching files when `FILE_UPLOAD=local`
+* Fix: Inline code blocks that includes doller sign are broken
 
 ## 2.4.1